+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.6] - 2020-03-17
+### Changed
+- UI changed, fixed un-matching state's name
+
 ## [0.0.5] - 2020-03-15
 ### Changed
 - UI changed, fixed un-matching state's name